Feature/enhancments

+++ app/src/main/java/app/example/rates/model/CurrencyRates.kt
@@ -3,6 +3,7 @@ package app.example.rates.model
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class CurrencyRates(
@@ -11,5 +12,5 @@ data class CurrencyRates(
     @Json(name = "date")
     val date: String,
     @Json(name = "rates")
-    val ratesMap: Map<String, Double>
+    val ratesMap: Map<String, BigDecimal>
 )
